Sections added to Dashboard

+++ app/src/main/java/pt/android/cvnotes/data/repository/SectionRepositoryImpl.kt
@@ -12,7 +12,7 @@ class SectionRepositoryImpl(
         return dao.getSections()
     }
 
-    override suspend fun getSectionById(id: Long): Section? {
+    override suspend fun getSectionById(id: Int): Section? {
         return dao.getSectionById(id)
     }
